+++ server/camcops_server/__init__.py
@@ -5,7 +5,7 @@ camcops_server/__init__.py
 
 ===============================================================================
 
-    Copyright (C) 2012-2019 Rudolf Cardinal (rudolf@pobox.com).
+    Copyright (C) 2012-2020 Rudolf Cardinal (rudolf@pobox.com).
 
     This file is part of CamCOPS.
 
@@ -27,3 +27,9 @@ camcops_server/__init__.py
 The mere existence of this file makes Python treat the directory as a package.
 
 """
+
+# Create camcops_server.__version__:
+# noinspection PyPep8Naming
+from camcops_server.cc_modules.cc_version_string import (
+    CAMCOPS_SERVER_VERSION_STRING as __version__
+)