+++ server/camcops_server/alembic/versions/0012_idnum_def_validation_method.py
@@ -5,7 +5,7 @@ camcops_server/alembic/versions/0012_idnum_def_validation_method.py
 
 ===============================================================================
 
-    Copyright (C) 2012-2019 Rudolf Cardinal (rudolf@pobox.com).
+    Copyright (C) 2012-2020 Rudolf Cardinal (rudolf@pobox.com).
 
     This file is part of CamCOPS.