+++ server/playing/sqlalchemy_comment_mysql_bug.py
@@ -5,7 +5,7 @@ playing/sqlalchemy_comment_mysql_bug.py
 
 ===============================================================================
 
-    Copyright (C) 2012-2019 Rudolf Cardinal (rudolf@pobox.com).
+    Copyright (C) 2012-2020 Rudolf Cardinal (rudolf@pobox.com).
 
     This file is part of CamCOPS.